completed unit test and refactored some code

+++ src/services/productPhotosServices.ts
@@ -12,7 +12,7 @@ export async function GetProductPhotosService(productId: string) {
 
     return photos;
   } catch (error) {
-    throw new AppError("can not get photos", 500);
+    throw error;
   }
 }
 
@@ -20,32 +20,34 @@ export async function SetDefaultProductPhotoService(
   photoId: string,
   isDefault: boolean
 ) {
-  const photo = await prisma.productPhotos.findUnique({
-    where: {
-      id: photoId,
-    },
-  });
-
-  if (!photo) throw new AppError("Product photo not found", 404);
-
-  const newPhoto = await prisma.$transaction(async (tx) => {
-    await tx.productPhotos.updateMany({
-      where: { productId: photo.productId },
-      data: { isDefault: false },
-    });
-
-    const updatedPhoto = await tx.productPhotos.update({
-      where: { id: photoId },
-      data: {
-        isDefault,
+  try {
+    const photo = await prisma.productPhotos.findUnique({
+      where: {
+        id: photoId,
       },
     });
-    return updatedPhoto;
-  });
 
-  if (!newPhoto) throw new AppError("can not set photo to default", 500);
+    if (!photo) throw new AppError("Product photo not found", 404);
 
-  return newPhoto;
+    const newPhoto = await prisma.$transaction(async (tx) => {
+      await tx.productPhotos.updateMany({
+        where: { productId: photo.productId },
+        data: { isDefault: false },
+      });
+
+      const updatedPhoto = await tx.productPhotos.update({
+        where: { id: photoId },
+        data: {
+          isDefault,
+        },
+      });
+      return updatedPhoto;
+    });
+
+    return newPhoto;
+  } catch (error) {
+    throw error;
+  }
 }
 
 export async function UpdateProductPhotoService(
@@ -56,6 +58,7 @@ export async function UpdateProductPhotoService(
     const existingProductPhoto = await prisma.productPhotos.findUnique({
       where: { id: photoId },
     });
+
     if (!existingProductPhoto) throw new AppError("Product not found", 404);
 
     let publicId = `products/product_${photoId}_${Date.now()}`; // Default for new images
@@ -85,7 +88,7 @@ export async function UpdateProductPhotoService(
 
     return imageUrl;
   } catch (error) {
-    throw new AppError("can not upload image", 500);
+    throw error;
   }
 }
 
@@ -93,38 +96,56 @@ export async function AddProductPhotoService(
   fileUri: string,
   productId: string
 ) {
-  const existingProduct = await prisma.products.findUnique({
-    where: { id: productId },
-  });
-  if (!existingProduct) throw new AppError("Product not found", 404);
+  try {
+    const existingProduct = await prisma.products.findUnique({
+      where: { id: productId },
+    });
 
-  let publicId = `products/product_${productId}_${Date.now()}`; // Default for new images
+    if (!existingProduct) throw new AppError("Product not found", 404);
 
-  // Upload the image to Cloudinary
-  const uploadResult = await cloudinary.uploader.upload(fileUri, {
-    public_id: publicId,
-    overwrite: true,
-    folder: "products",
-  });
+    let publicId = `products/product_${productId}_${Date.now()}`; // Default for new images
 
-  const imageUrl = uploadResult.secure_url;
+    // Upload the image to Cloudinary
+    const uploadResult = await cloudinary.uploader.upload(fileUri, {
+      public_id: publicId,
+      overwrite: true,
+      folder: "products",
+    });
 
-  const newPhoto = await prisma.productPhotos.create({
-    data: { imageUrl, productId, updatedAt: new Date() },
-  });
+    const imageUrl = uploadResult.secure_url;
 
-  if (!newPhoto) throw new AppError("can not upload new photo", 500);
+    const newPhoto = await prisma.productPhotos.create({
+      data: { imageUrl, productId, updatedAt: new Date() },
+    });
 
-  return newPhoto;
+    if (!newPhoto) throw new AppError("can not upload new photo", 500);
+
+    return newPhoto;
+  } catch (error) {
+    throw error;
+  }
 }
 
 export async function DeleteProductPhotoService(photoId: string) {
   try {
+    const photo = await prisma.productPhotos.findUnique({
+      where: { id: photoId },
+    });
+
+    if (!photo) throw new AppError("Product photo to delete not found", 404);
+
+    if (photo.imageUrl) {
+      const publicId = getPublicIdFromUrl(photo.imageUrl);
+      if (publicId) {
+        await cloudinary.uploader.destroy(publicId);
+      }
+    }
+
     const deletedPhoto = await prisma.productPhotos.delete({
       where: { id: photoId },
     });
     return deletedPhoto;
   } catch (error) {
-    throw new AppError("can not delete photo", 500);
+    throw error;
   }
 }
